requirements.txt
numpy==1.18.3
pygame==1.9.6
PyOpenGL==3.1.5
